Drivers: Rename ad400x to pulsar-adc
The ad400x hdl is used for a broader set of parts like the ad798x series.
This is called the "pulsar_adc" in the HDL project.

Follow the same naming convention in No-Os driver/project.
