--- doc/sphinx/source/drivers/ad400x.rst
@@ -1,1 +1,0 @@
-.. include:: ../../../../drivers/adc/ad400x/README.rst
