+++ doc/sphinx/source/drivers/pulsar_adc.rst
@@ -1,0 +1,1 @@
+.. include:: ../../../../drivers/adc/pulsar_adc/README.rst
