--- doc/sphinx/source/projects/ad400x-fmcz.rst
@@ -1,1 +1,0 @@
-.. include:: ../../../../projects/ad400x-fmcz/README.rst
