+++ doc/sphinx/source/projects/pulsar-adc.rst
@@ -1,0 +1,1 @@
+.. include:: ../../../../projects/pulsar-adc/README.rst
