+++ drivers/adc/pulsar_adc/iio_pulsar_adc.h
@@ -1,6 +1,6 @@
 /***************************************************************************//**
- *   @file   main.c
- *   @brief  Main file for STM32 platform of ad400x-fmcz project.
+ *   @file   iio_pulsar_adc.h
+ *   @brief  Header file for PULSAR_ADC IIO interface
  *   @author Axel Haslam (ahaslam@baylibre.com)
 ********************************************************************************
  * Copyright 2024(c) Analog Devices, Inc.
@@ -37,43 +37,49 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
 
+#ifndef _PULSAR_ADC_IIO_H_
+#define _PULSAR_ADC_IIO_H_
+
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-#include "platform_includes.h"
-#include "common_data.h"
-#include "no_os_error.h"
+#include <stdint.h>
 
-#ifdef BASIC_EXAMPLE
-#include "basic_example.h"
-#elif defined(IIO_EXAMPLE)
-#include "iio_example.h"
-#endif
+#include "iio.h"
+#include "iio_types.h"
+#include "pulsar_adc.h"
 
-/***************************************************************************//**
- * @brief Main function execution for STM32 platform.
- *
- * @return ret - Result of the enabled examples execution.
-*******************************************************************************/
-int main()
-{
-	int ret = -EINVAL;
+/******************************************************************************/
+/********************** Public/Extern Declarations ****************************/
+/******************************************************************************/
+/**
+ * @struct pulsar_adc_iio_dev
+ * @brief pulsar_adc IIO device structure
+ */
+struct pulsar_adc_iio_dev {
+	/** pulsar_adc driver handler */
+	struct pulsar_adc_dev *pulsar_adc_dev;
+	/** Generic IIO device handler */
+	struct iio_device *iio_dev;
+	/* Voltage reference */
+	uint32_t ref_voltage_mv;
+	/* scan type */
+	struct scan_type scan_type;
+};
 
-	stm32_init();
-#ifdef BASIC_EXAMPLE
-	struct no_os_uart_desc *uart;
+/**
+ * @struct pulsar_adc_iio_init_param
+ * @brief pulsar_adc IIO initialization structure
+ */
+struct pulsar_adc_iio_init_param {
+	struct pulsar_adc_init_param *init_param;
+	/* Voltage reference */
+	uint32_t ref_voltage_mv;
+};
 
-	ret = no_os_uart_init(&uart, &ad400x_uart_ip);
-	if (ret)
-		return ret;
-
-	no_os_uart_stdio(uart);
-
-	ret = basic_example_main();
-#elif defined(IIO_EXAMPLE)
-	ret = iio_example_main();
-#else
-#error At least one example has to be selected using y value in Makefile.
-#endif
-	return ret;
-}
+/** Allocate memory for pulsar_adc IIO handler. */
+int pulsar_adc_iio_init(struct pulsar_adc_iio_dev **dev,
+			struct pulsar_adc_iio_init_param *iio_init_param);
+/** Free memory allocated by iio_pulsar_adc_init(). */
+int pulsar_adc_iio_remove(struct pulsar_adc_iio_dev *dev);
+#endif /* __PULSAR_ADC_IIO_H__ */
