+++ projects/pulsar-adc/src/platform/stm32/main.c
@@ -1,7 +1,6 @@
 /***************************************************************************//**
- *   @file   parameters.h
- *   @brief  Definitions specific to STM32 platform used by eval-ad400x
- *           project.
+ *   @file   main.c
+ *   @brief  Main file for STM32 platform of pulsar-adc project.
  *   @author Axel Haslam (ahaslam@baylibre.com)
 ********************************************************************************
  * Copyright 2024(c) Analog Devices, Inc.
@@ -37,61 +36,44 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
-#ifndef __PARAMETERS_H__
-#define __PARAMETERS_H__
 
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-#include "stm32_hal.h"
-#include "stm32_irq.h"
-#include "stm32_gpio_irq.h"
-#include "stm32_spi.h"
-#include "stm32_gpio.h"
-#include "stm32_uart.h"
-#include "stm32_uart_stdio.h"
+#include "platform_includes.h"
+#include "common_data.h"
+#include "no_os_error.h"
 
-/******************************************************************************/
-/********************** Macros and Constants Definitions **********************/
-/******************************************************************************/
-extern UART_HandleTypeDef huart5;
-#define SAMPLES_PER_CHANNEL_PLATFORM 2000
-#define MAX_SIZE_BASE_ADDR     (SAMPLES_PER_CHANNEL_PLATFORM * sizeof(uint32_t))
-extern uint8_t in_buff[];
-
-#define ADC_DDR_BASEADDR	in_buff
-
-#ifdef IIO_SUPPORT
-#define INTC_DEVICE_ID 0
+#ifdef BASIC_EXAMPLE
+#include "basic_example.h"
+#elif defined(IIO_EXAMPLE)
+#include "iio_example.h"
 #endif
 
-#define UART_IRQ_ID		UART5_IRQn
+/***************************************************************************//**
+ * @brief Main function execution for STM32 platform.
+ *
+ * @return ret - Result of the enabled examples execution.
+*******************************************************************************/
+int main()
+{
+	int ret = -EINVAL;
 
-#define UART_DEVICE_ID		5
-#define UART_BAUDRATE		230400
-#define UART_EXTRA		&ad400x_uart_extra_ip
-#define UART_OPS		&stm32_uart_ops
+	stm32_init();
+#ifdef BASIC_EXAMPLE
+	struct no_os_uart_desc *uart;
 
-/*
- * spi cs is not used, but current no-os mandates
- * a cs to be specifed. So we need to add some
- * dummy gpio value
- */
-#define SPI_DEVICE_ID		1
-#define SPI_BAUDRATE		20000000
-#define SPI_CS			14
-#define SPI_CS_PORT		GPIO_PORT_A
-#define SPI_OPS			&stm32_spi_ops
-#define SPI_EXTRA		&ad400x_spi_extra_ip
+	ret = no_os_uart_init(&uart, &pulsar_adc_uart_ip);
+	if (ret)
+		return ret;
 
-#define GPIO_CNV		15
-#define GPIO_OPS		&stm32_gpio_ops
-#define GPIO_CNV_PORT		GPIO_PORT_A
+	no_os_uart_stdio(uart);
 
-#define GPIO_PORT_A		0
-
-#define AD400X_ADC_REF_VOLTAGE 5000
-extern struct stm32_uart_init_param ad400x_uart_extra_ip;
-extern struct stm32_spi_init_param ad400x_spi_extra_ip;
-
-#endif /* __PARAMETERS_H__ */
+	ret = basic_example_main();
+#elif defined(IIO_EXAMPLE)
+	ret = iio_example_main();
+#else
+#error At least one example has to be selected using y value in Makefile.
+#endif
+	return ret;
+}
