+++ projects/pulsar-adc/src/platform/xilinx/main.c
@@ -1,6 +1,6 @@
 /***************************************************************************//**
- *   @file   iio_example.h
- *   @brief  IIO example header for eval-ad400x project
+ *   @file   main.c
+ *   @brief  Main file for xilinx platform of pulsar-adc project.
  *   @author Axel Haslam (ahaslam@baylibre.com)
 ********************************************************************************
  * Copyright 2024(c) Analog Devices, Inc.
@@ -36,12 +36,40 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
-#ifndef __IIO_EXAMPLE_H__
-#define __IIO_EXAMPLE_H__
 
 /******************************************************************************/
-/************************ Functions Declarations ******************************/
+/***************************** Include Files **********************************/
 /******************************************************************************/
-int iio_example_main();
+#include "platform_includes.h"
+#include "common_data.h"
+#include "no_os_error.h"
 
-#endif /* __IIO_EXAMPLE_H__ */
+#ifdef BASIC_EXAMPLE
+#include "basic_example.h"
+#elif defined(IIO_EXAMPLE)
+#include "iio_example.h"
+#endif
+
+/***************************************************************************//**
+ * @brief Main function execution for xilinx platform.
+ *
+ * @return ret - Result of the enabled examples execution.
+*******************************************************************************/
+int main()
+{
+	int ret = -EINVAL;
+
+	/* Enable the instruction cache. */
+	Xil_ICacheEnable();
+	/* Enable the data cache. */
+	Xil_DCacheEnable();
+
+#ifdef BASIC_EXAMPLE
+	ret = basic_example_main();
+#elif defined(IIO_EXAMPLE)
+	ret = iio_example_main();
+#else
+#error At least one example has to be selected using y value in Makefile.
+#endif
+	return ret;
+}
